Get credible intervals for a, R, b

Get credible intervals for a, R, b

Usage

credible_prep(res, params = c("a", "R", "b"), labels = FALSE)

Arguments

res

output data frame from ebase

params

character vector indicating which parameters to plot, one to any of a, R, or b (default all)

labels

logical indicating of parameter labels are output as an expression for parsing in plot facets, default FALSE

Value

A data frame

Details

This function gets 95% credible intervals (2.5th to 97.5th percentiles, approximate posterior distributions) for a, R, and/or b using the output from ebase. The function is used in credible_plot, but is provided as a separate function for convenience.

Examples

library(dplyr)
library(lubridate)
library(doParallel)

# get four days of data
dat <- exdat %>% 
  filter(month(exdat$DateTimeStamp) == 6 & day(exdat$DateTimeStamp) %in% 1:4)

##
# run ebase with defaults, parallel

# setup parallel backend
cl <- makeCluster(2)
registerDoParallel(cl)

res <- ebase(dat, interval = 900, H = 1.85, progress = TRUE, n.chains = 2)

stopCluster(cl)

# get credible intervals
credible_prep(res)
#> # A tibble: 12 × 6
#> # Groups:   grp [4]
#>    Date         grp var       mean      lo      hi
#>    <date>     <dbl> <fct>    <dbl>   <dbl>   <dbl>
#>  1 2012-06-01     1 a       0.0964  0.0570  0.135 
#>  2 2012-06-01     1 R     127.     NA      NA     
#>  3 2012-06-01     1 b       0.283   0.104   0.455 
#>  4 2012-06-02     2 a       0.103   0.0750  0.131 
#>  5 2012-06-02     2 R     165.     NA      NA     
#>  6 2012-06-02     2 b       0.267   0.0875  0.443 
#>  7 2012-06-03     3 a       0.0709  0.0458  0.0967
#>  8 2012-06-03     3 R     125.     NA      NA     
#>  9 2012-06-03     3 b       0.301   0.124   0.461 
#> 10 2012-06-04     4 a       0.0845  0.0602  0.110 
#> 11 2012-06-04     4 R     135.     NA      NA     
#> 12 2012-06-04     4 b       0.262   0.0896  0.437