Estuarine Bayesian Single-station Estimation method for ecosystem metabolism

Estuarine Bayesian Single-station Estimation method for ecosystem metabolism

Usage

ebase(
  dat,
  Z,
  interval,
  ndays = 1,
  aprior = c(4, 2),
  rprior = c(300, 150),
  bprior = c(0.251, 0.125),
  bmax = 0.502,
  nogas = FALSE,
  doave = TRUE,
  maxinterp = 43200/interval,
  n.iter = 10000,
  update.chains = TRUE,
  n.burnin = n.iter * 0.5,
  n.chains = 3,
  n.thin = 10,
  progress = NULL,
  model_file = NULL
)

Arguments

dat

input data frame

Z

numeric as single value for water column depth (m) or vector equal in length to number of rows in dat

interval

timestep interval in seconds

ndays

numeric for number of days in dat for optimizing the metabolic equation, see details

aprior

numeric vector of length two indicating the mean and standard deviation for the prior distribution of the a parameter, see details

rprior

numeric vector of length two indicating the mean and standard deviation for the prior distribution of the R parameter, see details

bprior

numeric vector of length two indicating the mean and standard deviation for the prior distribution of the b parameter, see details

bmax

numeric value for the upper limit on the prior distribution for bprior, set as twice the default value of the mean

nogas

logical indicating if gas exchange is not included in the metabolic model, see details

doave

logical indicating if the average dissolved oxygen concentration is used as the starting value for the estimation (default), otherwise the first observation will be used if FALSE, see details

maxinterp

numeric value for minimum number of continuous observations that must not be interpolated within a group defined by ndays to assign as NA in output, see details

n.iter

number of MCMC iterations, passed to jags

update.chains

logical to run metab_update if chains do not converge

n.burnin

number of MCMC chains to delete, passed to jags

n.chains

number of MCMC chains to run, passed to jags

n.thin

number of nth iterations to save for each chain, passed to jags

progress

character string of path where progress is saved to as 'log.txt', use NULL to suppress (default)

model_file

NULL to use the model file included with the package or a path to a model text file can be used

Value

A data frame with metabolic estimates for areal gross production (P, O2 mmol m-2 d-1), respiration (R, O2 mmol m-2 d-1), and gas exchange (D, O2 mmol m-2 d-1, positive values as ingassing, negative values as outgassing). Additional parameters estimated by the model that are returned include a and b. The a parameter is a constant that represents the primary production per quantum of light with units of (mmol m-2 d-1)/(W m-2) and is used to estimate gross production (Grace et al., 2015). The b parameter is a constant used to estimate gas exchange in (cm hr-1)/(m2 s-2) (provided as 0.251 in eqn. 4 in Wanninkhof 2014). Observed dissolved oxygen (DO_obs, mmol m-3), modeled dissolved oxygen (DO_mod, mmol m-3), and delta dissolved oxygen of the modeled results (dDO, mmol m-3 d-1) are also returned. Note that delta dissolved oxygen is a daily rate.

95% credible intervals for a, b, and R are also returned in the corresponding columns alo, ahi, blo, bhi, Rlo, and Rhi, for the 2.5th and 97.5th percentile estimates for each parameter, respectively. These values indicate the interval within which there is a 95% probability that the true parameter is in this range. Note that all values for these parameters are repeated across rows, although only one estimate for each is returned based on the number of days defined by ndays.

Model fit can also be assessed using the converge and rsq columns. The values in these columns apply to each group in the grp column as specified with the ndays argument. The converge column indicates "Check convergence" or "Fine" if the JAGS estimate converged at that iteration (repeated across rows for the group). The n.chains argument can be increased if convergence is not achieved. Similarly, the rsq column shows the r-squared values of the linear fit between the modeled and observed dissolved oxygen (repeated across rows for the group). These values can also be viewed with fit_plot.

The nogas argument can be set to TRUE to exclude gas exchange from the metabolic estimates. This will force the prior distribution for b as mean 0 and standard deviation approximately 0.

Details

Required input data are time series for dissolved oxygen (mg L-1), water temperature (C), salinity (psu), total PAR (W m-2), and wind speed (m s-1). See the exdat example data file for a representation of the required data. Data are typically from continuously monitored water quality and weather parameters are hourly of sub-hourly time steps. Oxygen concentrations are converted to mmol/m3 prior to metabolic estimation. Water column depth is also required. This can be supplied as a single value or a vector of length equal to the number of rows in dat.

The metabolic estimates are based on a mass balance equation in Grace et al. 2015 with the gas exchange estimate from Wanninkhof 2004. It is similar to that provided by the BASEmetab R package at https://github.com/dgiling/BASEmetab, with modifications to estimate different parameters. The equation optimized in the JAGS model is:

$$ Z\frac{dC_d}{dt} = aPAR - R + bU_{10}^2\left(\frac{Sc}{600} \right)^{-0.5} \left(C_{Sat} - C_d \right )$$

More simply:

$$ Z\frac{dC_d}{dt} = P - R + D$$

Gross production is provided by aPAR, respiration is provided by R, and gas exchange is provided by the remainder. The likelihood of the parameters a, R, and b given the observed data are estimated from the JAGS model using prior distributions shown in the model file. At each time step, the change in oxygen concentration between time steps is calculated from the equation using model inputs and parameter guesses, and then a finite difference approximation is used to estimate modeled oxygen concentration. The first modeled value starts at the mean oxygen concentration for all measurements in the optimization period. The estimated concentration at each time step is also returned for comparison to observed as one measure of model performance.

The prior distributions for the a, R, and b parameters are defined in the model file included with the package as normal distributions with mean and standard deviations provided by the aprior, rprior, and bprior arguments. The default values were chosen based on approximate values from national syntheses of metabolic estimates. An additional constraint is that all the prior distributions are truncated to be positive values as required by the core metabolism equation above. The upper limit for b is set as two times 0.251, as given in eqn. 4 in Wanninkhof 2014. Units for each parameter are (mmol m-2 d-1)/(W m-2) for a, mmol m-2 d-1 for R, and (cm hr-1)/(m2 s-2) for b.

The ndays argument defines the model optimization period as the number of days that are used for optimizing the above mass balance equation. By default, this is done each day, i.e., ndays= 1 such that a loop is used that applies the model equation to observations within each day, evaluated iteratively from the first observation in a day to the last. Individual parameter estimates for a, R, and b are then returned for each day. However, more days can be used to estimate the unknown parameters, such that the loop can be evaluated for every ndays specified by the argument. The ndays argument will separate the input data into groups of consecutive days, where each group has a total number of days equal to ndays. The final block may not include the complete number of days specified by ndays if the number of unique dates in the input data includes a remainder when divided by ndays, e.g., if seven days are in the input data and ndays = 5, there will be two groups where the first has five days and the second has two days. The output data from ebase includes a column that specifies the grouping that was used based on ndays.

Missing values in the input data are also interpolated prior to estimating metabolism. It is the responsibility of the user to verify that these interpolated values are not wildly inaccurate. Missing values are linearly interpolated between non-missing values at the time step specified by the value in interval. This works well for small gaps, but can easily create inaccurate values at gaps larger than a few hours. The interp_plot function can be used to visually assess the interpolated values. Records at the start or end of the input time series that do not include a full day are also removed. A warning is returned to the console if gaps are found or dangling records are found.

The maxinterp argument specifies a minimum number of observations that must not be interpolated within groups defined by ndays that are assigned NA in the output (except Date and DateTimeStamp). Groups with continuous rows of interpolated values with length longer than this argument are assigned NA. The default value is half a day, i.e., 43200 seconds divided by the value in interval.

The doave argument can be used to define which dissolved oxygen value is used as the starting point in the Bayesian estimation for the optimization period. The default setting (doave = TRUE) will use the average of all the dissolved oxygen values in the optimization period defined by ndays. For example, the average of all dissolved oxygen values in each 24 hour period will be used if doave = TRUE and ndays = 1. The first dissolved oxygen observation of the time series in the optimization period will be used as the starting point if doave = F. In this case, the simulated dissolved oxygen time series will always start at the first observed dissolved oxygen value for each optimization period.

References

Grace, M.R., Giling, D.P., Hladyz, S., Caron, V., Thompson, R.M., Nally, R.M., 2015. Fast processing of diel oxygen curves: Estimating stream metabolism with BASE (BAyesian Single-station Estimation). Limnology and Oceanography: Methods 13, e10011. https://doi.org/10.1002/lom3.10011

Wanninkhof, R., 2014. Relationship between wind speed and gas exchange over the ocean revisited. Limnology and Oceanography: Methods 12, 351–362. https://doi.org/10.4319/lom.2014.12.351

Examples

# get one day of data
dat <- exdat[as.Date(exdat$DateTimeStamp, tz = 'America/Jamaica') == as.Date('2012-06-01'), ]

# run ebase, use more chains and iterations for a better fit, update.chains as T
ebase(dat, interval = 900, Z = 1.85, n.chains = 2, n.iter = 50, 
 update.chains = FALSE)
#> Warning: executing %dopar% sequentially: no parallel backend registered
#> Loading required package: rjags
#> Loading required package: coda
#> Linked to JAGS 4.3.0
#> Loaded modules: basemod,bugs
#> 
#> Attaching package: ‘R2jags’
#> The following object is masked from ‘package:coda’:
#> 
#>     traceplot
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
#>          DateTimeStamp       Date grp    Z  DO_obs   DO_mod DO_modlo DO_modhi
#> 1  2012-06-01 00:00:00 2012-06-01   1 1.85 140.625 166.1132 166.1126 166.1138
#> 2  2012-06-01 00:15:00 2012-06-01   1 1.85 159.375 165.3935 165.3220 165.5432
#> 3  2012-06-01 00:30:00 2012-06-01   1 1.85 150.000 164.7099 164.5634 165.0194
#> 4  2012-06-01 00:45:00 2012-06-01   1 1.85 150.000 164.0037 163.7845 164.4663
#> 5  2012-06-01 01:00:00 2012-06-01   1 1.85 150.000 163.2883 162.9972 163.9008
#> 6  2012-06-01 01:15:00 2012-06-01   1 1.85 156.250 162.5434 162.1829 163.2974
#> 7  2012-06-01 01:30:00 2012-06-01   1 1.85 168.750 161.8078 161.3773 162.7054
#> 8  2012-06-01 01:45:00 2012-06-01   1 1.85 168.750 161.0710 160.5707 162.1115
#> 9  2012-06-01 02:00:00 2012-06-01   1 1.85 165.625 160.3363 159.7663 161.5200
#> 10 2012-06-01 02:15:00 2012-06-01   1 1.85 165.625 159.6247 158.9835 160.9572
#> 11 2012-06-01 02:30:00 2012-06-01   1 1.85 150.000 158.8924 158.1816 160.3680
#> 12 2012-06-01 02:45:00 2012-06-01   1 1.85 159.375 158.2528 157.4662 159.8946
#> 13 2012-06-01 03:00:00 2012-06-01   1 1.85 131.250 157.5133 156.6577 159.2957
#> 14 2012-06-01 03:15:00 2012-06-01   1 1.85 118.750 156.7306 155.8091 158.6429
#> 15 2012-06-01 03:30:00 2012-06-01   1 1.85 121.875 155.9822 154.9926 158.0324
#> 16 2012-06-01 03:45:00 2012-06-01   1 1.85 115.625 155.2355 154.1778 157.4239
#> 17 2012-06-01 04:00:00 2012-06-01   1 1.85 118.750 154.5368 153.4081 156.8745
#> 18 2012-06-01 04:15:00 2012-06-01   1 1.85 131.250 153.7835 152.5872 156.2575
#> 19 2012-06-01 04:30:00 2012-06-01   1 1.85 156.250 153.1179 151.8489 155.7483
#> 20 2012-06-01 04:45:00 2012-06-01   1 1.85 162.500 152.4391 151.0983 155.2225
#> 21 2012-06-01 05:00:00 2012-06-01   1 1.85 134.375 151.7478 150.3362 154.6809
#> 22 2012-06-01 05:15:00 2012-06-01   1 1.85 150.000 151.0578 149.5754 154.1406
#> 23 2012-06-01 05:30:00 2012-06-01   1 1.85 150.000 150.6448 149.0761 153.9393
#> 24 2012-06-01 05:45:00 2012-06-01   1 1.85 146.875 150.1248 148.4763 153.6060
#> 25 2012-06-01 06:00:00 2012-06-01   1 1.85 150.000 149.4518 147.7335 153.0811
#> 26 2012-06-01 06:15:00 2012-06-01   1 1.85 150.000 148.7725 146.9885 152.5382
#> 27 2012-06-01 06:30:00 2012-06-01   1 1.85 137.500 148.1075 146.2624 151.9978
#> 28 2012-06-01 06:45:00 2012-06-01   1 1.85 140.625 147.4426 145.5408 151.4447
#> 29 2012-06-01 07:00:00 2012-06-01   1 1.85 150.000 146.8060 144.8515 150.9108
#> 30 2012-06-01 07:15:00 2012-06-01   1 1.85 153.125 146.2360 144.2331 150.4359
#> 31 2012-06-01 07:30:00 2012-06-01   1 1.85 159.375 145.9500 143.9422 150.1416
#> 32 2012-06-01 07:45:00 2012-06-01   1 1.85 168.750 146.1790 144.1881 150.3368
#> 33 2012-06-01 08:00:00 2012-06-01   1 1.85 168.750 146.6067 144.6472 150.7040
#> 34 2012-06-01 08:15:00 2012-06-01   1 1.85 168.750 146.5962 144.5981 150.8006
#> 35 2012-06-01 08:30:00 2012-06-01   1 1.85 162.500 147.0068 145.0675 151.0649
#> 36 2012-06-01 08:45:00 2012-06-01   1 1.85 165.625 147.5822 145.7171 151.4479
#> 37 2012-06-01 09:00:00 2012-06-01   1 1.85 168.750 148.4423 146.6538 152.1143
#> 38 2012-06-01 09:15:00 2012-06-01   1 1.85 171.875 149.4842 147.7818 152.9424
#> 39 2012-06-01 09:30:00 2012-06-01   1 1.85 171.875 151.3224 149.5144 154.7564
#> 40 2012-06-01 09:45:00 2012-06-01   1 1.85 168.750 153.3084 151.0879 156.7723
#> 41 2012-06-01 10:00:00 2012-06-01   1 1.85 168.750 153.5643 151.0514 157.3020
#> 42 2012-06-01 10:15:00 2012-06-01   1 1.85 175.000 153.7756 151.1353 157.5758
#> 43 2012-06-01 10:30:00 2012-06-01   1 1.85 178.125 154.1028 151.4211 157.8240
#> 44 2012-06-01 10:45:00 2012-06-01   1 1.85 178.125 155.4089 152.7609 158.7262
#> 45 2012-06-01 11:00:00 2012-06-01   1 1.85 178.125 156.6951 154.1425 159.5301
#> 46 2012-06-01 11:15:00 2012-06-01   1 1.85 175.000 158.0054 155.5797 160.3095
#> 47 2012-06-01 11:30:00 2012-06-01   1 1.85 162.500 159.3458 157.0526 161.1047
#> 48 2012-06-01 11:45:00 2012-06-01   1 1.85 165.625 160.7277 158.5634 162.4647
#> 49 2012-06-01 12:00:00 2012-06-01   1 1.85 159.375 162.2002 160.1510 164.1497
#> 50 2012-06-01 12:15:00 2012-06-01   1 1.85 159.375 164.0740 162.0282 166.2938
#> 51 2012-06-01 12:30:00 2012-06-01   1 1.85 162.500 165.6557 163.5178 168.1274
#> 52 2012-06-01 12:45:00 2012-06-01   1 1.85 165.625 167.2160 164.9952 169.9521
#> 53 2012-06-01 13:00:00 2012-06-01   1 1.85 153.125 168.7745 166.5460 171.7842
#> 54 2012-06-01 13:15:00 2012-06-01   1 1.85 140.625 169.6355 167.3747 172.8035
#> 55 2012-06-01 13:30:00 2012-06-01   1 1.85 153.125 171.0881 168.7735 174.4969
#> 56 2012-06-01 13:45:00 2012-06-01   1 1.85 153.125 172.4370 170.0801 176.0689
#> 57 2012-06-01 14:00:00 2012-06-01   1 1.85 146.875 174.0277 171.4462 177.9163
#> 58 2012-06-01 14:15:00 2012-06-01   1 1.85 143.750 175.6071 172.6233 179.7471
#> 59 2012-06-01 14:30:00 2012-06-01   1 1.85 131.250 177.2261 173.7999 181.6251
#> 60 2012-06-01 14:45:00 2012-06-01   1 1.85 168.750 178.4056 174.7343 182.9782
#> 61 2012-06-01 15:00:00 2012-06-01   1 1.85 175.000 179.2437 175.4161 183.9356
#> 62 2012-06-01 15:15:00 2012-06-01   1 1.85 150.000 180.4956 176.4631 185.3485
#> 63 2012-06-01 15:30:00 2012-06-01   1 1.85 190.625 182.2572 177.9474 187.3224
#> 64 2012-06-01 15:45:00 2012-06-01   1 1.85 181.250 183.6345 179.0204 188.8803
#> 65 2012-06-01 16:00:00 2012-06-01   1 1.85 200.000 184.6039 179.8325 189.9585
#> 66 2012-06-01 16:15:00 2012-06-01   1 1.85 200.000 185.6549 180.7628 191.1090
#> 67 2012-06-01 16:30:00 2012-06-01   1 1.85 196.875 186.5096 181.4786 192.0542
#> 68 2012-06-01 16:45:00 2012-06-01   1 1.85 203.125 186.9396 181.8542 192.5207
#> 69 2012-06-01 17:00:00 2012-06-01   1 1.85 206.250 187.4099 182.2663 193.0298
#> 70 2012-06-01 17:15:00 2012-06-01   1 1.85 206.250 188.1034 182.8538 193.7891
#> 71 2012-06-01 17:30:00 2012-06-01   1 1.85 206.250 188.6005 183.2892 194.3237
#> 72 2012-06-01 17:45:00 2012-06-01   1 1.85 209.375 188.6086 183.3211 194.3162
#> 73 2012-06-01 18:00:00 2012-06-01   1 1.85 209.375 188.5256 183.2894 194.2035
#> 74 2012-06-01 18:15:00 2012-06-01   1 1.85 206.250 188.6583 183.4490 194.3238
#> 75 2012-06-01 18:30:00 2012-06-01   1 1.85 206.250 188.5367 183.3448 194.1820
#> 76 2012-06-01 18:45:00 2012-06-01   1 1.85 203.125 188.1724 183.0369 193.7744
#> 77 2012-06-01 19:00:00 2012-06-01   1 1.85 203.125 187.7099 182.6760 193.2488
#> 78 2012-06-01 19:15:00 2012-06-01   1 1.85 203.125 187.1669 182.2311 192.6415
#> 79 2012-06-01 19:30:00 2012-06-01   1 1.85 203.125 186.5392 181.7461 191.9323
#> 80 2012-06-01 19:45:00 2012-06-01   1 1.85 196.875 185.8228 181.1774 191.1287
#> 81 2012-06-01 20:00:00 2012-06-01   1 1.85 190.625 185.0887 180.6045 190.3033
#> 82 2012-06-01 20:15:00 2012-06-01   1 1.85 193.750 184.3134 179.9768 189.4386
#> 83 2012-06-01 20:30:00 2012-06-01   1 1.85 196.875 183.5295 179.3350 188.5668
#> 84 2012-06-01 20:45:00 2012-06-01   1 1.85 184.375 182.7656 178.7225 187.7145
#> 85 2012-06-01 21:00:00 2012-06-01   1 1.85 175.000 182.0113 178.1232 186.8720
#> 86 2012-06-01 21:15:00 2012-06-01   1 1.85 175.000 181.2683 177.5394 186.0415
#> 87 2012-06-01 21:30:00 2012-06-01   1 1.85 171.875 180.5435 176.9813 185.2295
#> 88 2012-06-01 21:45:00 2012-06-01   1 1.85 103.125 179.8355 176.4471 184.4340
#> 89 2012-06-01 22:00:00 2012-06-01   1 1.85 150.000 179.1411 175.9311 183.6531
#> 90 2012-06-01 22:15:00 2012-06-01   1 1.85 162.500 178.4414 175.4043 182.8700
#> 91 2012-06-01 22:30:00 2012-06-01   1 1.85 131.250 177.7537 174.8949 182.0980
#> 92 2012-06-01 22:45:00 2012-06-01   1 1.85 156.250 177.0818 174.2971 181.3449
#> 93 2012-06-01 23:00:00 2012-06-01   1 1.85 181.250 176.3702 173.5785 180.5507
#> 94 2012-06-01 23:15:00 2012-06-01   1 1.85 184.375 175.6662 172.8640 179.7671
#> 95 2012-06-01 23:30:00 2012-06-01   1 1.85 175.000 174.9048 172.1102 178.9240
#> 96 2012-06-01 23:45:00 2012-06-01   1 1.85 162.500 174.1385 171.3529 178.0762
#>            dDO          converge       rsq         a       alo      ahi
#> 1           NA Check convergence 0.3337845        NA        NA       NA
#> 2  -69.0845445 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 3  -65.6341784 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 4  -67.7908401 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 5  -68.6818929 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 6  -71.5087325 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 7  -70.6181749 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 8  -70.7349228 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 9  -70.5251236 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 10 -68.3168691 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 11 -70.2988904 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 12 -61.3975880 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 13 -70.9991111 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 14 -75.1375021 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 15 -71.8483676 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 16 -71.6808220 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 17 -67.0754081 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 18 -72.3179934 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 19 -63.9000233 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 20 -65.1629775 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 21 -66.3663545 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 22 -66.2378607 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 23 -39.6493899 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 24 -49.9162819 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 25 -64.6112712 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 26 -65.2113953 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 27 -63.8356181 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 28 -63.8299368 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 29 -61.1172233 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 30 -54.7187001 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 31 -27.4601952 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 32  21.9862620 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 33  41.0567243 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 34  -1.0084089 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 35  39.4182628 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 36  55.2455934 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 37  82.5644871 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 38 100.0281093 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 39 176.4600609 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 40 190.6623332 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 41  24.5619293 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 42  20.2840622 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 43  31.4164510 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 44 125.3838967 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 45 123.4716720 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 46 125.7862796 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 47 128.6860914 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 48 132.6619019 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 49 141.3552692 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 50 179.8844632 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 51 151.8478700 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 52 149.7855768 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 53 149.6143351 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 54  82.6617505 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 55 139.4506920 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 56 129.4872855 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 57 152.7074503 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 58 151.6205561 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 59 155.4270051 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 60 113.2374808 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 61  80.4529014 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 62 120.1851855 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 63 169.1147005 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 64 132.2185448 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 65  93.0623577 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 66 100.8942002 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 67  82.0507096 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 68  41.2762453 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 69  45.1545472 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 70  66.5740846 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 71  47.7220995 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 72   0.7781415 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 73  -7.9639075 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 74  12.7371610 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 75 -11.6712709 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 76 -34.9806159 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 77 -44.3937295 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 78 -52.1283680 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 79 -60.2605873 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 80 -68.7799893 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 81 -70.4723335 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 82 -74.4235308 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 83 -75.2547301 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 84 -73.3334668 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 85 -72.4149286 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 86 -71.3313843 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 87 -69.5795626 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 88 -67.9672993 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 89 -66.6643811 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 90 -67.1718490 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 91 -66.0131950 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 92 -64.5039541 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 93 -68.3105879 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 94 -67.5864931 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 95 -73.0927775 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#> 96 -73.5705521 Check convergence 0.3337845 0.9122096 0.6136428 1.056313
#>            b       blo       bhi          P         Plo        Phi        R
#> 1         NA        NA        NA         NA          NA         NA       NA
#> 2  0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 3  0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 4  0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 5  0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 6  0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 7  0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 8  0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 9  0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 10 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 11 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 12 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 13 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 14 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 15 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 16 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 17 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 18 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 19 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 20 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 21 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 22 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 23 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 24 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 25 0.3482113 0.2079293 0.4644247   1.366794   0.9194415   1.582709 146.9914
#> 26 0.3482113 0.2079293 0.4644247   4.761734   3.2032155   5.513954 146.9914
#> 27 0.3482113 0.2079293 0.4644247   9.589604   6.4509200  11.104490 146.9914
#> 28 0.3482113 0.2079293 0.4644247  13.778167   9.2685632  15.954727 146.9914
#> 29 0.3482113 0.2079293 0.4644247  18.760352  12.6200757  21.723956 146.9914
#> 30 0.3482113 0.2079293 0.4644247  25.969088  17.4693880  30.071469 146.9914
#> 31 0.3482113 0.2079293 0.4644247  80.596763  54.2173875  93.328770 146.9914
#> 32 0.3482113 0.2079293 0.4644247 125.260068  84.2623621 145.047613 146.9914
#> 33 0.3482113 0.2079293 0.4644247 148.164892  99.6704216 171.570751 146.9914
#> 34 0.3482113 0.2079293 0.4644247  61.329375  41.2562287  71.017680 146.9914
#> 35 0.3482113 0.2079293 0.4644247 177.176200 119.1863083 205.165024 146.9914
#> 36 0.3482113 0.2079293 0.4644247 207.201580 139.3843613 239.933565 146.9914
#> 37 0.3482113 0.2079293 0.4644247 230.855937 155.2966307 267.324640 146.9914
#> 38 0.3482113 0.2079293 0.4644247 252.107381 169.5924626 291.933210 146.9914
#> 39 0.3482113 0.2079293 0.4644247 259.338162 174.4566046 300.306251 146.9914
#> 40 0.3482113 0.2079293 0.4644247 254.686653 171.3275376 294.919935 146.9914
#> 41 0.3482113 0.2079293 0.4644247  37.873424  25.4774266  43.856353 146.9914
#> 42 0.3482113 0.2079293 0.4644247  98.629628  66.3480830 114.210317 146.9914
#> 43 0.3482113 0.2079293 0.4644247 151.846418 102.1469817 175.833854 146.9914
#> 44 0.3482113 0.2079293 0.4644247 331.844386 223.2314917 384.266406 146.9914
#> 45 0.3482113 0.2079293 0.4644247 354.661030 238.5802324 410.687434 146.9914
#> 46 0.3482113 0.2079293 0.4644247 371.591641 249.9694429 430.292602 146.9914
#> 47 0.3482113 0.2079293 0.4644247 378.447657 254.5814800 438.231674 146.9914
#> 48 0.3482113 0.2079293 0.4644247 383.319617 257.8588440 443.873265 146.9914
#> 49 0.3482113 0.2079293 0.4644247 391.432201 263.3161740 453.267409 146.9914
#> 50 0.3482113 0.2079293 0.4644247 406.312621 273.3262223 470.498514 146.9914
#> 51 0.3482113 0.2079293 0.4644247 329.485564 221.6447137 381.534957 146.9914
#> 52 0.3482113 0.2079293 0.4644247 338.149275 227.4727863 391.567289 146.9914
#> 53 0.3482113 0.2079293 0.4644247 375.008626 252.2680466 434.249374 146.9914
#> 54 0.3482113 0.2079293 0.4644247 247.433827 166.4485659 286.521367 146.9914
#> 55 0.3482113 0.2079293 0.4644247 333.828442 224.5661648 386.563886 146.9914
#> 56 0.3482113 0.2079293 0.4644247 322.144557 216.7064232 373.034278 146.9914
#> 57 0.3482113 0.2079293 0.4644247 364.272679 245.0459821 421.817451 146.9914
#> 58 0.3482113 0.2079293 0.4644247 365.088347 245.5946811 422.761971 146.9914
#> 59 0.3482113 0.2079293 0.4644247 380.431713 255.9161531 440.529155 146.9914
#> 60 0.3482113 0.2079293 0.4644247 290.950788 195.7223961 336.912776 146.9914
#> 61 0.3482113 0.2079293 0.4644247 236.741969 159.2561609 274.140499 146.9914
#> 62 0.3482113 0.2079293 0.4644247 292.560078 196.8049642 338.776289 146.9914
#> 63 0.3482113 0.2079293 0.4644247 365.088347 245.5946811 422.761971 146.9914
#> 64 0.3482113 0.2079293 0.4644247 332.726189 223.8246798 385.287508 146.9914
#> 65 0.3482113 0.2079293 0.4644247 258.412270 173.8337571 299.234093 146.9914
#> 66 0.3482113 0.2079293 0.4644247 261.454489 175.8802559 302.756897 146.9914
#> 67 0.3482113 0.2079293 0.4644247 243.818436 164.0164950 282.334847 146.9914
#> 68 0.3482113 0.2079293 0.4644247 179.733427 120.9065536 208.126221 146.9914
#> 69 0.3482113 0.2079293 0.4644247 185.597415 124.8512542 214.916553 146.9914
#> 70 0.3482113 0.2079293 0.4644247 220.847476 148.5639463 255.735126 146.9914
#> 71 0.3482113 0.2079293 0.4644247 190.755961 128.3214042 220.890003 146.9914
#> 72 0.3482113 0.2079293 0.4644247 117.456114  79.0126478 136.010856 146.9914
#> 73 0.3482113 0.2079293 0.4644247 101.693892  68.4094115 117.758648 146.9914
#> 74 0.3482113 0.2079293 0.4644247 131.983813  88.7854209 152.833520 146.9914
#> 75 0.3482113 0.2079293 0.4644247 103.435452  69.5809579 119.775326 146.9914
#> 76 0.3482113 0.2079293 0.4644247  66.972912  45.0526322  77.552736 146.9914
#> 77 0.3482113 0.2079293 0.4644247  46.933947  31.5724338  54.348182 146.9914
#> 78 0.3482113 0.2079293 0.4644247  37.586838  25.2846405  43.524495 146.9914
#> 79 0.3482113 0.2079293 0.4644247  18.583991  12.5014381  21.519736 146.9914
#> 80 0.3482113 0.2079293 0.4644247   7.076466   4.7603341   8.194348 146.9914
#> 81 0.3482113 0.2079293 0.4644247   2.072236   1.3939919   2.399591 146.9914
#> 82 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 83 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 84 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 85 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 86 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 87 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 88 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 89 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 90 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 91 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 92 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 93 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 94 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 95 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#> 96 0.3482113 0.2079293 0.4644247   0.000000   0.0000000   0.000000 146.9914
#>         Rlo      Rhi          D        Dlo        Dhi
#> 1        NA       NA         NA         NA         NA
#> 2  125.7794 161.9542  19.184992  11.456044  25.588138
#> 3  125.7794 161.9542  25.568170  15.283172  33.999706
#> 4  125.7794 161.9542  21.578345  12.916049  28.604665
#> 5  125.7794 161.9542  19.929898  11.942202  26.343637
#> 6  125.7794 161.9542  14.700245   8.816710  19.378750
#> 7  125.7794 161.9542  16.347776   9.811937  21.489896
#> 8  125.7794 161.9542  16.131792   9.690760  21.141481
#> 9  125.7794 161.9542  16.519921   9.930156  21.599198
#> 10 125.7794 161.9542  20.605192  12.394603  26.872279
#> 11 125.7794 161.9542  16.938452  10.197996  22.035419
#> 12 125.7794 161.9542  33.405862  20.123851  43.369190
#> 13 125.7794 161.9542  15.643044   9.437889  20.253777
#> 14 125.7794 161.9542   7.987021   4.820844  10.322325
#> 15 125.7794 161.9542  14.071920   8.496539  18.139004
#> 16 125.7794 161.9542  14.381879   8.686655  18.506487
#> 17 125.7794 161.9542  22.901895  13.837838  29.418386
#> 18 125.7794 161.9542  13.203112   7.983256  16.931801
#> 19 125.7794 161.9542  28.776357  17.404846  36.841164
#> 20 125.7794 161.9542  26.439891  16.006751  33.793166
#> 21 125.7794 161.9542  24.213644  14.673097  30.883015
#> 22 125.7794 161.9542  24.451357  14.829521  31.124038
#> 23 125.7794 161.9542  73.640028  44.676096  93.700092
#> 24 125.7794 161.9542  54.646278  33.246091  69.360132
#> 25 125.7794 161.9542  26.093754  15.908978  33.031759
#> 26 125.7794 161.9542  21.588584  13.169302  27.296986
#> 27 125.7794 161.9542  19.305902  11.784664  24.364516
#> 28 125.7794 161.9542  15.127850   9.238406  19.063125
#> 29 125.7794 161.9542  15.164185   9.263508  19.080005
#> 30 125.7794 161.9542  19.792716  12.090840  24.888858
#> 31 125.7794 161.9542  15.593275   9.526820  19.597742
#> 32 125.7794 161.9542  62.405917  38.120730  78.445652
#> 33 125.7794 161.9542  74.781448  45.751676  93.997373
#> 34 125.7794 161.9542  83.796468  51.353649 105.416893
#> 35 125.7794 161.9542  42.738986  26.263335  53.681244
#> 36 125.7794 161.9542  41.994167  25.823194  52.802206
#> 37 125.7794 161.9542  68.879764  42.348851  86.878299
#> 38 125.7794 161.9542  79.936021  49.225314 100.988909
#> 39 125.7794 161.9542 214.104350 132.095415 271.109245
#> 40 125.7794 161.9542 245.030063 152.361035 310.206309
#> 41 125.7794 161.9542 154.557545  97.067558 195.373578
#> 42 125.7794 161.9542  85.887287  54.273217 108.132663
#> 43 125.7794 161.9542  53.265416  33.751255  67.003289
#> 44 125.7794 161.9542  47.107222  29.851795  59.387971
#> 45 125.7794 161.9542  20.752963  13.168174  26.291640
#> 46 125.7794 161.9542   8.104376   5.135325  10.346402
#> 47 125.7794 161.9542   6.613012   4.182870   8.514495
#> 48 125.7794 161.9542   9.096302   5.738452  11.823195
#> 49 125.7794 161.9542  17.066447  10.740810  22.399000
#> 50 125.7794 161.9542  73.465036  46.161045  97.392758
#> 51 125.7794 161.9542  98.424395  61.942609 131.722581
#> 52 125.7794 161.9542  85.945442  54.299494 115.770093
#> 53 125.7794 161.9542  48.769293  30.900006  66.192921
#> 54 125.7794 161.9542  52.481812  33.268589  72.016400
#> 55 125.7794 161.9542  71.146738  45.109676  98.130375
#> 56 125.7794 161.9542  64.398321  40.929482  89.641770
#> 57 125.7794 161.9542  65.227504  41.528771  91.629004
#> 58 125.7794 161.9542  62.401082  39.790210  88.675694
#> 59 125.7794 161.9542  54.099646  34.545761  77.833370
#> 60 125.7794 161.9542  65.529951  41.886169  95.666245
#> 61 125.7794 161.9542  59.087298  37.934926  87.301785
#> 62 125.7794 161.9542  76.773915  49.468617 114.402461
#> 63 125.7794 161.9542  94.765249  61.363911 142.817612
#> 64 125.7794 161.9542  58.869519  38.574760  90.845533
#> 65 125.7794 161.9542  60.744492  39.785368  94.792888
#> 66 125.7794 161.9542  72.191181  47.380180 113.555126
#> 67 125.7794 161.9542  54.966777  36.240962  87.263760
#> 68 125.7794 161.9542  43.619026  28.908883  70.172776
#> 69 125.7794 161.9542  44.929897  29.825400  72.591760
#> 70 125.7794 161.9542  49.305980  32.716088  79.746598
#> 71 125.7794 161.9542  44.521323  29.706466  72.976756
#> 72 125.7794 161.9542  30.974847  20.735349  51.170102
#> 73 125.7794 161.9542  30.564279  20.493300  50.567878
#> 74 125.7794 161.9542  38.571334  25.897152  63.802299
#> 75 125.7794 161.9542  21.964096  14.815259  36.648751
#> 76 125.7794 161.9542  15.304348  10.295891  25.376662
#> 77 125.7794 161.9542  17.929053  12.037363  29.547728
#> 78 125.7794 161.9542  12.967080   8.670688  21.137058
#> 79 125.7794 161.9542  16.925322  11.192210  27.002144
#> 80 125.7794 161.9542  12.671953   8.361323  20.022324
#> 81 125.7794 161.9542  14.545346   9.559458  22.708235
#> 82 125.7794 161.9542   9.307868   6.112307  14.419606
#> 83 125.7794 161.9542   7.770149   5.086926  11.920978
#> 84 125.7794 161.9542  11.324486   7.383315  17.191550
#> 85 125.7794 161.9542  13.023782   8.468334  19.595733
#> 86 125.7794 161.9542  15.028339   9.739544  22.394713
#> 87 125.7794 161.9542  18.269209  11.831649  27.053588
#> 88 125.7794 161.9542  21.251896  13.788213  31.355624
#> 89 125.7794 161.9542  23.662295  15.335747  34.655486
#> 90 125.7794 161.9542  22.723479  14.619778  32.821695
#> 91 125.7794 161.9542  24.866989  16.099923  35.944392
#> 92 125.7794 161.9542  27.659085  17.794543  39.517930
#> 93 125.7794 161.9542  20.616812  13.345937  29.445249
#> 94 125.7794 161.9542  21.956387  14.069231  30.956877
#> 95 125.7794 161.9542  11.769761   7.562453  16.554990
#> 96 125.7794 161.9542  10.885878   6.975261  15.229995