Prepare data for ebase

Prepare data for ebase

Usage

ebase_prep(dat, H, interval, ndays = 1, interp = TRUE, maxgap = 1e+06)

Arguments

dat

input data frame

H

numeric as single value for water column depth (m) or vector equal in length to number of rows in dat

interval

timestep interval in seconds

ndays

numeric for number of days in dat for optimizing the metabolic equation, see details

interp

logical indicating if linear interpolation is used to fill all data gaps, only applies if timestep specified by interval is not uniform

maxgap

numeric vector indicating maximum gap size to interpolate where size is number of records at the value specified by interval, e.g., if interval = 900 and maxgap = 4, gaps up to and including one hour will be linearly interpolated

Value

A data frame with additional columns required for ebase. If multiple time steps are identified, the number of rows in data frame is expanded based on the time step define by interval. Numeric values in the expanded rows will be interpolated if interp = TRUE, otherwise they will remain as NA values.

Details

Checks if all columns are present by matching those in exdat, converts dissolved oxygen from mg/L to mmol/m3, calculates the Schmidt number (unitless) from water temp (C) and salinity (psu), and calculates dissolved oxygen equilibrium concentration (mmol/m3) from salinity and temperature

The ndays argument defines the number of days that are used for optimizing the above mass balance equation. By default, this is done each day, i.e., ndays= 1 such that a loop is used that applies the model equation to observations within each day, evaluated iteratively from the first observation in a day to the last. Individual parameter estimates for a, r, and b are then returned for each day. However, more days can be used to estimate the unknown parameters, such that the loop can be evaluated for every ndays specified by the argument. The ndays argument will separate the input data into groups of consecutive days, where each group has a total number of days equal to ndays. The final block may not include the complete number of days specified by ndays if the number of unique dates in the input data includes a remainder when divided by ndays, e.g., if seven days are in the input data and ndays = 5, there will be two groups where the first has five days and the second has two days. The output data from ebase includes a column that specifies the grouping that was used based on ndays.

Records at the start or end of the input time series that do not include a full day are removed for conformance with the core model equation. A warning is returned to the console if these records are found.

Examples

dat <- ebase_prep(exdat, H = 1.85, interval = 900)
#> More than one time step observed: 900, 778500, 260100
#> Interpolating missing values to interval
head(dat)
#> # A tibble: 6 × 12
#>   Date       DateTimeStamp       isinterp DO_obs DO_sat     H  Temp   Sal   PAR
#>   <date>     <dttm>              <lgl>     <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 2012-02-23 2012-02-23 00:00:00 FALSE      275    266.  1.85  16.4  23       0
#> 2 2012-02-23 2012-02-23 00:15:00 FALSE      275    266.  1.85  16.4  22.8     0
#> 3 2012-02-23 2012-02-23 00:30:00 FALSE      275    266.  1.85  16.4  22.7     0
#> 4 2012-02-23 2012-02-23 00:45:00 FALSE      275    266.  1.85  16.4  22.9     0
#> 5 2012-02-23 2012-02-23 01:00:00 FALSE      272.   266.  1.85  16.4  22.7     0
#> 6 2012-02-23 2012-02-23 01:15:00 FALSE      266.   265.  1.85  16.4  23.4     0
#> # … with 3 more variables: WSpd <dbl>, sc <dbl>, grp <dbl>